prepare for 40mhz support

+++ itl80211/openbsd/net80211/ieee80211_ioctl.h
@@ -333,7 +333,7 @@ struct ieee80211_nodereq {
 
 	/* Channel and rates */
 	u_int16_t	nr_channel;			/* last channel */
-	u_int16_t	nr_chan_flags;			/* channel flags */
+	u_int32_t	nr_chan_flags;		/* channel flags */
 	u_int8_t	nr_nrates;			/* rate count */
 	u_int8_t	nr_rates[IEEE80211_RATE_MAXSIZE];	/* rate set */
